test: add unit test
