+++ packages/utilities/src/lib/tryParseJSON.ts
@@ -2,7 +2,7 @@
  * Try parse a stringified JSON string.
  * @param value The value to parse
  */
-export function tryParseJSON(value: string): object | string {
+export function tryParseJSON(value: string): object | string | number {
 	try {
 		return JSON.parse(value);
 	} catch (err) {
@@ -15,6 +15,6 @@ export function tryParseJSON(value: string): object | string {
  * @param value The value to parse
  * @deprecated Use {@link tryParseJSON} instead
  */
-export function tryParse(value: string): object | string {
+export function tryParse(value: string): object | string | number {
 	return tryParseJSON(value);
 }
