+++ packages/utilities/tests/tryParseJSON.test.ts
@@ -1,15 +1,15 @@
-import { tryParse } from '../src';
+import { tryParseJSON } from '../src';
 
-describe('tryParse', () => {
+describe('tryParseJSON', () => {
 	test('GIVEN basic THEN returns expected', () => {
 		const source = '{"a":4,"b":6}';
 		const expected = { a: 4, b: 6 };
-		expect(tryParse(source)).toEqual(expected);
+		expect(tryParseJSON(source)).toEqual(expected);
 	});
 
 	test('GIVEN invalid THEN returns expected', () => {
 		const source = '{"a":4,"b:6}';
 		const expected = '{"a":4,"b:6}';
-		expect(tryParse(source)).toEqual(expected);
+		expect(tryParseJSON(source)).toEqual(expected);
 	});
 });
